Trade-off: too many/too few context-switches
In theory, spliting the HTML file list exactly by number of
cores, creates a perfect workload for each of the cores.
Sadly, that's not the case - because the .html files
may or may not have 'href's; meaning that one task may
finish quite quickly, and have nothing to do.

On the other end, if we split by too much, we waste
time in the workers context-switching (to read from the
consumer_Q) instead of processing HTML hrefs!